Rename index.ts to cli.test
It better reflects it's purpose, and index.ts is more
canonically used for lib exports.
