+++ bin/cli.js
@@ -1,0 +1,6 @@
+#!/usr/bin/env node
+
+import cli from "../dist/cli.js";
+
+// eslint-disable-next-line no-undef
+await cli(process.argv);
