--- bin/invoke.js
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-
-import index from "../dist/index.js";
-
-// eslint-disable-next-line no-undef
-await index(process.argv);
